% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newValue.R
\name{newValue}
\alias{newValue}
\title{Initializing the constant parameters of the allometric equations.}
\usage{
newValue()
}
\value{
Returns a list, with all parameters indexable, Numeric.
}
\description{
From the parameters stored in code, this function builds and returns a list, which is from top to bottom nested
provinces, plant species, equations, and vectors of two parameters identified by plant organs.
}
\examples{
valueSet = newValue()
}
\references{
Zhou et al. _Carbon stocks of forest ecosystems in China: biomass equations_. Beijing: Science Press (2018): 48-52.
}
