% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WCalculate.R
\name{WCalculate}
\alias{WCalculate}
\title{Finding the constant parameters and substituting all parameters into the equation.}
\usage{
WCalculate(
  ifIncludeH = FALSE,
  valueSet,
  prov = 0,
  spe = 0,
  D = 0,
  H = 0,
  resChoose = 0
)
}
\arguments{
\item{ifIncludeH}{Value of selected allometric equation (includes the parameter of tree height or not) to be used, Logical.}

\item{valueSet}{List of all parameters, Numeric.}

\item{prov}{Vector of numerically indexed province names, Numeric.}

\item{spe}{Vector of numerically indexed plant species, Numeric.}

\item{D}{Vector of the DBH of trees, Numeric.}

\item{H}{Vector of the height of trees, Numeric.}

\item{resChoose}{Value of selected components of tree biomass to be calculated and returned, Numeric. To be chosen between:
- `1`: Biomass of stem
- `2`: Biomass of branch
- `3`: Biomass of leaf
- `4`: Biomass of root}
}
\value{
Returns a vecotr of calculation results of tree biomass.
}
\description{
From the numerically indexed province names and plant species, this function finds the constant parameters of
the corresponding equation in the value set of all parameters, and then gives the calculation results the equation returned.
}
\examples{
# Load a database
data(ChinaBiomassData)
# Initialize a value set of all parameters
vSet = newValue()
WCalculate(ifIncludeH = TRUE, valueSet = vSet, prov = c(1, 2, 3), spe = c(1, 2, 3), D = ChinaBiomassData$DBH, H = ChinaBiomassData$H, resChoose = 1)
}
