% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WEquation.R
\name{WEquation}
\alias{WEquation}
\title{Calculating the equation.}
\usage{
WEquation(ifIncludeH = FALSE, a = 0, b = 0, D = 0, H = 0)
}
\arguments{
\item{ifIncludeH}{Value of selected allometric equation (includes the parameter of tree height or not) to be used, Logical.}

\item{a}{Vector of the constant parameters 'a' of the equation, Numeric.}

\item{b}{Vector of the constant parameters 'b' of the equation, Numeric.}

\item{D}{Vector of the DBH of trees, Numeric.}

\item{H}{Vector of the height of trees, Numeric.}
}
\value{
Returns a vecotr of calculation results of the equation.
}
\description{
Based on the selected allometric equation, this function gives the calculation results of tree biomass
by substituting all constant and variable parameters into the equation.
}
\examples{
# Load a database
data(ChinaBiomassData)
# Get constant parameters from a value set
va = valueSet(a)
vb = valueSet(b)
WEquation(ifIncludeH = TRUE, a = va, b = vb, D = ChinaBiomassData$DBH, H = ChinaBiomassData$H)
}
